Use `uint16_t` for sprite widths

+++ Source/engine/load_cel.cpp
@@ -4,12 +4,12 @@
 
 namespace devilution {
 
-OwnedCelSprite LoadCel(const char *pszName, int width)
+OwnedCelSprite LoadCel(const char *pszName, uint16_t width)
 {
 	return OwnedCelSprite(LoadFileInMem(pszName), width);
 }
 
-OwnedCelSprite LoadCel(const char *pszName, const int *widths)
+OwnedCelSprite LoadCel(const char *pszName, const uint16_t *widths)
 {
 	return OwnedCelSprite(LoadFileInMem(pszName), widths);
 }
