+++ Source/engine/load_cel.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 #include "engine/cel_sprite.hpp"
 
 namespace devilution {
@@ -7,7 +9,7 @@ namespace devilution {
 /**
  * @brief Loads a Cel sprite and sets its width
  */
-OwnedCelSprite LoadCel(const char *pszName, int width);
-OwnedCelSprite LoadCel(const char *pszName, const int *widths);
+OwnedCelSprite LoadCel(const char *pszName, uint16_t width);
+OwnedCelSprite LoadCel(const char *pszName, const uint16_t *widths);
 
 } // namespace devilution
